refactor: move components and services into src
