+++ AGENTS.md
@@ -4,12 +4,12 @@ This repository contains a React Native (Expo) application. New contributors can
 
 ## Structure
 
-- `components/` – React Native UI components.
+- `src/components/` – React Native UI components.
 - `src/` – application logic, translations under `src/locales`, and Jest tests in `src/__tests__`.
-- `services/` – networking helpers and domain services.
-- `domain/` – pure domain logic shared by services and components.
+- `src/services/` – networking helpers and domain services.
+- `src/domain/` – pure domain logic shared by services and components.
 - `assets/` – static images and resources.
-- `App.js` – application entry point; renders map, HUD, and menus.
+- `App.tsx` – application entry point; renders map, HUD, and menus.
 
 The app fetches routing data and traffic light information from remote APIs. Tests live beside the code they verify, and coverage outputs under `coverage/`.
 
@@ -44,7 +44,7 @@ Tests should pass and formatting hooks should run on the changed files.
 ## For New Contributors
 
 - Start the app with `npm start` to launch the Expo dev server.
-- UI components live in `components/` and rely on hooks for state.
+- UI components live in `src/components/` and rely on hooks for state.
 - Core logic and translations live under `src/` (`src/i18n.ts` wires up `src/locales`).
-- Network and domain helpers are in `services/` and `domain/`.
+- Network and domain helpers are in `src/services/` and `src/domain/`.
 - Jest tests reside in `src/__tests__`; add tests alongside new features.
