+++ package.json
@@ -8,7 +8,7 @@
     "ios": "expo run:ios",
     "web": "expo start --web",
     "test": "jest",
-    "lint": "eslint src components services"
+    "lint": "eslint src"
   },
   "dependencies": {
     "@react-native-async-storage/async-storage": "^1.24.0",
