+++ src/components/CameraScreen.tsx
