+++ src/services/trafficLightDetector.ts
