+++ src/supabase.test.ts
@@ -18,10 +18,10 @@ describe('fetchLightsAndCycles error handling', () => {
   });
 
   it('returns error when fetching lights fails', async () => {
-    const logger = await import('../services/logger');
+    const logger = await import('./services/logger');
     const logSpy = jest.spyOn(logger, 'log').mockResolvedValue(undefined);
     const { supabaseService, supabase } = await import(
-      '../services/supabase'
+      './services/supabase'
     );
     const { fetchLightsAndCycles } = supabaseService;
     const originalFrom = supabase.from;
@@ -43,10 +43,10 @@ describe('fetchLightsAndCycles error handling', () => {
   });
 
   it('returns error when fetching cycles fails', async () => {
-    const logger = await import('../services/logger');
+    const logger = await import('./services/logger');
     const logSpy = jest.spyOn(logger, 'log').mockResolvedValue(undefined);
     const { supabaseService, supabase } = await import(
-      '../services/supabase'
+      './services/supabase'
     );
     const { fetchLightsAndCycles } = supabaseService;
     const originalFrom = supabase.from;
